Move static-env imports
